C tests: deduplicate some test mains
Deduplicate standard yaep_parse_grammar() type tests.
