+++ test/C/common.h
@@ -5,6 +5,8 @@
 #include<stdio.h>
 #include<stdlib.h>
 
+#include"yaep.h"
+
 static void *
 test_parse_alloc (int size)
 {
@@ -40,6 +42,7 @@ test_syntax_error (int err_tok_num, void *err_tok_attr,
 }
 
 static const char *input;
+static const char *description;
 
 /* The following function imported by YAEP (see comments in the interface file). */
 static int
@@ -56,6 +59,33 @@ test_read_token (void **attr)
     {
       return -1;
     }
+}
+
+static void
+test_standard_parse (void)
+{
+  struct grammar *g;
+  struct yaep_tree_node *root;
+  int ambiguous_p;
+
+  if ((g = yaep_create_grammar ()) == NULL)
+    {
+      fprintf (stderr, "yaep_create_grammar: No memory\n");
+      exit (1);
+    }
+  if (yaep_parse_grammar (g, 1, description) != 0)
+    {
+      fprintf (stderr, "%s\n", yaep_error_message (g));
+      exit (1);
+    }
+  if (yaep_parse
+      (g, test_read_token, test_syntax_error, test_parse_alloc,
+       test_parse_free, &root, &ambiguous_p))
+    {
+      fprintf (stderr, "yaep parse: %s\n", yaep_error_message (g));
+      exit (1);
+    }
+  yaep_free_grammar (g);
 }
 
 #endif
