+++ test/C/test27.c
@@ -1,8 +1,6 @@
-#include<stdio.h>
 #include <stdlib.h>
 
 #include"common.h"
-#include "yaep.h"
 
 static const char *input = "a+a*(a*+a)";
 
@@ -21,24 +19,6 @@ static const char *description =
 
 main ()
 {
-  struct grammar *g;
-  struct yaep_tree_node *root;
-  int ambiguous_p;
-
-  if ((g = yaep_create_grammar ()) == NULL)
-    {
-      fprintf (stderr, "yaep_create_grammar: No memory\n");
-      exit (1);
-    }
-  if (yaep_parse_grammar (g, 1, description) != 0)
-    {
-      fprintf (stderr, "%s\n", yaep_error_message (g));
-      exit (1);
-    }
-  if ( yaep_parse( g, test_read_token, test_syntax_error, test_parse_alloc, test_parse_free, &root, &ambiguous_p ) ) {
-      fprintf (stderr, "yaep parse: %s\n", yaep_error_message (g));
-      exit (1);
-    }
-  yaep_free_grammar (g);
+  test_standard_parse ();
   exit (0);
 }
